calendar month starting day correction + tilt effect + header
There is now a cool tilt hover effect on cards, using the react-parallax-tilt library. There is also a "Harker Lunch Menu" header on the top of the page. When the user selects another month on the calendar selector, it doesn't change the currently selected month, and the days on the month are now correctly shown as the 1st of the month doesn't always start on Sunday (which is what was happening before.